requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.4
SQLAlchemy==2.0.27
alembic==1.13.1
python-multipart==0.0.9
pytest==8.1.1
httpx==0.27.0
pymysql==1.1.0
python-dotenv==1.0.1
cryptography==41.0.5
